added selling stocks without closing positions

+++ server/reactivestocks/portfolio/views.py
@@ -1,9 +1,11 @@
+from os import times
+import time
+from turtle import pos
 from django.http import JsonResponse
 from rest_framework.response import Response
 from rest_framework import status
-from piechart.models import Portfolio
-from .models import Position, Stock
-from .serializer import PositionSerializer
+from .models import Position, Stock, Transaction, Portfolio
+from .serializer import PositionSerializer, TransactionSerializer
 from follow.serializer import StockSerializer
 import requests
 from follow.views import update
@@ -32,6 +34,7 @@ class PortfolioView(APIView):
         except Portfolio.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
         except Exception as e:
+            print(e)
             return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
@@ -54,21 +57,20 @@ class OpenPositionView(APIView):
                     stock.users.add(request.user)
                     stock.save()
                 else:
-                    print(serializer.errors)
                     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
             if stock is None:
                 return Response({"error": "Stock could not be created or retrieved"}, status=status.HTTP_400_BAD_REQUEST)
             position = Position.objects.filter(
                 stock=stock, portfolio=portfolio, user=request.user).first()
-            if position:
-                position.average_price = (position.average_price * position.quantity +
-                                          float(data['average_price']) * float(data['quantity'])) / (position.quantity + float(data['quantity']))
-                position.quantity += float(data['quantity'])
-                position.save()
-            else:
-                position = Position(stock=stock, quantity=float(
-                    data['quantity']), average_price=float(data['average_price']), timestamp=data['date'], user=request.user, portfolio=portfolio)
-                position.save()
+            if position is None:
+                position = Position.objects.create(
+                    stock=stock, portfolio=portfolio, user=request.user, quantity=data['quantity'], average_price=data['average_price'])
+
+            Transaction.objects.create(
+                position=position, quantity=data['quantity'], average_price=data['average_price'], timestamp=data['timestamp'])
+
+            position.update()
+            position.save()
             return JsonResponse(PositionSerializer(position).data, status=status.HTTP_201_CREATED)
         except Portfolio.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -78,14 +80,25 @@ class OpenPositionView(APIView):
 
 
 class ClosePositionView(APIView):
-    permission_classes = (IsAuthenticated, )
 
-    def delete(self, request, pk):
+    def post(self, request, pk):
         try:
+            data = request.data
+            if 'quantity' not in data:
+                return Response({"error": "Quantity is required"}, status=status.HTTP_400_BAD_REQUEST)
+
             position = Position.objects.get(pk=pk)
-        except Position.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        stock = position.stock
-        stock.users.remove(request.user)
-        position.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            if position.quantity < float(data['quantity']):
+                return Response({"error": "Not enough shares to sell"}, status=status.HTTP_400_BAD_REQUEST)
+
+            position.save()
+            Transaction.objects.create(
+                position=position, quantity=-float(data['quantity']), average_price=data['average_price'], timestamp=None)
+            position.update()
+            position.save()
+            if position.quantity == 0:
+                position.delete()
+            return JsonResponse(PositionSerializer(position).data, status=status.HTTP_201_CREATED)
+        except Exception as e:
+            print(e)
+            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
